{{cookiecutter.project_slug}}/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/{{cookiecutter.project_slug}}"]

[project]
name = "{{cookiecutter.project_slug}}"
dynamic = ["version"]
description = "LangGraph + MCP agent template with TDD first."
authors = [{ name = "{{cookiecutter.author}}"}]
requires-python = ">=3.10"

dependencies = [
  "langgraph>=0.0.43",                 # graph runtime 
  "langchain-mcp-adapters>=0.1.7",     # MCP→LangGraph bridge 
  "mcp[cli]>=1.9.0",                 # talks to MCP servers 
  "a2a-protocol>=0.2.1",
  "pytest", "pytest-asyncio",          # TDD harness 
  "ruff", "black"                      # code-quality tools
]

[tool.ruff]
select = ["E","F","I","D","B"]
[tool.bandit]
skips = ["B101","B311"]
